+++ src/Resources/app/administration/src/module/swag-paypal/components/swag-paypal-settings-icon/index.js
@@ -1,0 +1,7 @@
+import template from './swag-paypal-settings-icon.html.twig';
+
+const { Component } = Shopware;
+
+Component.register('swag-paypal-settings-icon', {
+    template
+});